+++ app/src/main/java/com/miz/mizuu/fragments/MovieDiscoveryFragment.java
@@ -98,7 +98,7 @@ public class MovieDiscoveryFragment extends Fragment implements OnSharedPreferen
 		if (thumbnailSize.equals(getString(R.string.large))) 
 			mImageThumbSize = (int) (getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 1.33);
 		else if (thumbnailSize.equals(getString(R.string.normal))) 
-			mImageThumbSize = (int) (getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 1);
+			mImageThumbSize = getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 1;
 		else
 			mImageThumbSize = (int) (getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 0.75);
 		mImageThumbSpacing = getResources().getDimensionPixelSize(R.dimen.image_thumbnail_spacing);
@@ -304,7 +304,7 @@ public class MovieDiscoveryFragment extends Fragment implements OnSharedPreferen
 			if (thumbnailSize.equals(getString(R.string.large))) 
 				mImageThumbSize = (int) (getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 1.33);
 			else if (thumbnailSize.equals(getString(R.string.normal))) 
-				mImageThumbSize = (int) (getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 1);
+				mImageThumbSize = getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 1;
 			else
 				mImageThumbSize = (int) (getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 0.75);
