+++ app/src/main/java/com/miz/utils/ViewUtils.java
@@ -528,7 +528,7 @@ public class ViewUtils {
         if (thumbnailSize.equals(context.getString(R.string.large)))
             thumbSize = (int) (context.getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 1.33);
         else if (thumbnailSize.equals(context.getString(R.string.normal)))
-            thumbSize = (int) (context.getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 1);
+            thumbSize = (context.getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 1);
         else
             thumbSize = (int) (context.getResources().getDimensionPixelSize(R.dimen.image_thumbnail_size) * 0.75);
